requirements.txt
streamlit==1.32.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
openai==1.12.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
pytest==8.0.0 